+++ test/Integration/AdminApi/Action/AbstractActionTestCase.php
@@ -10,6 +10,7 @@ use Heptacom\HeptaConnect\Package\Shopware6\Http\AdminApi\Authentication\Contrac
 use Heptacom\HeptaConnect\Package\Shopware6\Http\AdminApi\Authentication\Contract\AuthenticatedHttpClientInterface;
 use Heptacom\HeptaConnect\Package\Shopware6\Http\AdminApi\Authentication\PortalNodeStorageAuthenticationStorage;
 use Heptacom\HeptaConnect\Package\Shopware6\Http\AdminApi\ErrorHandling\JsonResponseErrorHandler;
+use Heptacom\HeptaConnect\Package\Shopware6\Http\AdminApi\ErrorHandling\JsonResponseValidator\ServerErrorValidator;
 use Heptacom\HeptaConnect\Package\Shopware6\Support\JsonStreamUtility;
 use Heptacom\HeptaConnect\Package\Shopware6\Test\Support\Package\MemoryApiConfigurationStorage;
 use Heptacom\HeptaConnect\Package\Shopware6\Test\Support\Portal\MemoryPortalStorage;
@@ -45,6 +46,7 @@ abstract class AbstractActionTestCase extends TestCase
      * @template TActionClass of AbstractActionClient
      *
      * @param class-string<TActionClass> $actionClass
+     *
      * @return AbstractActionClient&TActionClass
      */
     protected function createAction(string $actionClass): AbstractActionClient
@@ -57,6 +59,7 @@ abstract class AbstractActionTestCase extends TestCase
             $this->createApiConfigurationStorage(),
             $jsonStreamUtility,
             new JsonResponseErrorHandler($jsonStreamUtility, [
+                new ServerErrorValidator(),
             ]),
         );
     }